chore: Regroup some adb commands
