+++ lib/adb.ts
@@ -23,6 +23,7 @@ import * as emuMethods from './tools/adb-emu-commands';
 import * as lockManagementCommands from './tools/lockmgmt';
 import * as keyboardCommands from './tools/keyboard-commands';
 import * as emuConstants from './tools/emu-constants';
+import * as deviceSettingsCommands from './tools/device-settings';
 
 
 export const DEFAULT_ADB_PORT = 5037;
@@ -144,7 +145,6 @@ export class ADB implements ADBOptions {
   initZipAlign = generalMethods.initZipAlign;
   initBundletool = generalMethods.initBundletool;
   getApiLevel = generalMethods.getApiLevel;
-  getPlatformVersion = generalMethods.getPlatformVersion;
   isDeviceConnected = generalMethods.isDeviceConnected;
   mkdir = generalMethods.mkdir;
   isValidClass = generalMethods.isValidClass;
@@ -159,35 +159,13 @@ export class ADB implements ADBOptions {
   getGrantedPermissions = generalMethods.getGrantedPermissions;
   getDeniedPermissions = generalMethods.getDeniedPermissions;
   getReqPermissions = generalMethods.getReqPermissions;
-  getLocationProviders = generalMethods.getLocationProviders;
-  toggleGPSLocationProvider = generalMethods.toggleGPSLocationProvider;
-  setHiddenApiPolicy = generalMethods.setHiddenApiPolicy;
-  setDefaultHiddenApiPolicy = generalMethods.setDefaultHiddenApiPolicy;
   stopAndClear = generalMethods.stopAndClear;
-  availableIMEs = generalMethods.availableIMEs;
-  enabledIMEs = generalMethods.enabledIMEs;
-  enableIME = generalMethods.enableIME;
-  disableIME = generalMethods.disableIME;
-  setIME = generalMethods.setIME;
-  defaultIME = generalMethods.defaultIME;
-  keyevent = generalMethods.keyevent;
-  inputText = generalMethods.inputText;
   clearTextField = generalMethods.clearTextField;
   lock = generalMethods.lock;
   back = generalMethods.back;
   goToHome = generalMethods.goToHome;
   getAdbPath = generalMethods.getAdbPath;
-  getScreenOrientation = generalMethods.getScreenOrientation;
   sendTelnetCommand = generalMethods.sendTelnetCommand;
-  isAirplaneModeOn = generalMethods.isAirplaneModeOn;
-  setAirplaneMode = generalMethods.setAirplaneMode;
-  setBluetoothOn = generalMethods.setBluetoothOn;
-  setNfcOn = generalMethods.setNfcOn;
-  broadcastAirplaneMode = generalMethods.broadcastAirplaneMode;
-  isWifiOn = generalMethods.isWifiOn;
-  isDataOn = generalMethods.isDataOn;
-  isAnimationOn = generalMethods.isAnimationOn;
-  setAnimationScale = generalMethods.setAnimationScale;
   rimraf = generalMethods.rimraf;
   push = generalMethods.push;
   pull = generalMethods.pull;
@@ -213,34 +191,12 @@ export class ADB implements ADBOptions {
   killProcessByPID = generalMethods.killProcessByPID;
   broadcastProcessEnd = generalMethods.broadcastProcessEnd;
   broadcast = generalMethods.broadcast;
-  getDeviceProperty = generalMethods.getDeviceProperty;
-  setDeviceProperty = generalMethods.setDeviceProperty;
-  getDeviceSysLanguage = generalMethods.getDeviceSysLanguage;
-  getDeviceSysCountry = generalMethods.getDeviceSysCountry;
-  getDeviceSysLocale = generalMethods.getDeviceSysLocale;
-  getDeviceProductLanguage = generalMethods.getDeviceProductLanguage;
-  getDeviceProductCountry = generalMethods.getDeviceProductCountry;
-  getDeviceProductLocale = generalMethods.getDeviceProductLocale;
-  getModel = generalMethods.getModel;
-  getManufacturer = generalMethods.getManufacturer;
-  getScreenSize = generalMethods.getScreenSize;
-  getScreenDensity = generalMethods.getScreenDensity;
-  setHttpProxy = generalMethods.setHttpProxy;
-  deleteHttpProxy = generalMethods.deleteHttpProxy;
-  setSetting = generalMethods.setSetting;
-  getSetting = generalMethods.getSetting;
   bugreport = generalMethods.bugreport;
   screenrecord = generalMethods.screenrecord;
-  runInImeContext = generalMethods.runInImeContext;
-  getTimeZone = generalMethods.getTimeZone;
   listFeatures = generalMethods.listFeatures;
   isStreamedInstallSupported = generalMethods.isStreamedInstallSupported;
   isIncrementalInstallSupported = generalMethods.isIncrementalInstallSupported;
-  getDeviceIdleWhitelist = generalMethods.getDeviceIdleWhitelist;
-  addToDeviceIdleWhitelist = generalMethods.addToDeviceIdleWhitelist;
   takeScreenshot = generalMethods.takeScreenshot;
-  setWifiState = generalMethods.setWifiState;
-  setDataState = generalMethods.setDataState;
   listPorts = generalMethods.listPorts;
 
   executeApksigner = apkSigningMethods.executeApksigner;
@@ -267,10 +223,6 @@ export class ADB implements ADBOptions {
   getApplicationInstallState = apkUtilsMethods.getApplicationInstallState;
   installOrUpgrade = apkUtilsMethods.installOrUpgrade;
   extractStringsFromApk = apkUtilsMethods.extractStringsFromApk;
-  getDeviceLanguage = apkUtilsMethods.getDeviceLanguage;
-  getDeviceCountry = apkUtilsMethods.getDeviceCountry;
-  getDeviceLocale = apkUtilsMethods.getDeviceLocale;
-  ensureCurrentLocale = apkUtilsMethods.ensureCurrentLocale;
   getApkInfo = apkUtilsMethods.getApkInfo;
   getPackageInfo = apkUtilsMethods.getPackageInfo;
   pullApk = apkUtilsMethods.pullApk;
@@ -278,6 +230,15 @@ export class ADB implements ADBOptions {
 
   hideKeyboard = keyboardCommands.hideKeyboard;
   isSoftKeyboardPresent = keyboardCommands.isSoftKeyboardPresent;
+  keyevent = keyboardCommands.keyevent;
+  availableIMEs = keyboardCommands.availableIMEs;
+  enabledIMEs = keyboardCommands.enabledIMEs;
+  enableIME = keyboardCommands.enableIME;
+  disableIME = keyboardCommands.disableIME;
+  setIME = keyboardCommands.setIME;
+  defaultIME = keyboardCommands.defaultIME;
+  inputText = keyboardCommands.inputText;
+  runInImeContext = keyboardCommands.runInImeContext;
 
   isLockManagementSupported = lockManagementCommands.isLockManagementSupported;
   verifyLockCredential = lockManagementCommands.verifyLockCredential;
@@ -369,4 +330,45 @@ export class ADB implements ADBOptions {
   GSM_SIGNAL_STRENGTHS = emuConstants.GSM_SIGNAL_STRENGTHS;
   NETWORK_SPEED = emuConstants.NETWORK_SPEED;
   SENSORS = emuConstants.SENSORS;
+
+  getDeviceProperty = deviceSettingsCommands.getDeviceProperty;
+  setDeviceProperty = deviceSettingsCommands.setDeviceProperty;
+  getDeviceSysLanguage = deviceSettingsCommands.getDeviceSysLanguage;
+  getDeviceSysCountry = deviceSettingsCommands.getDeviceSysCountry;
+  getDeviceSysLocale = deviceSettingsCommands.getDeviceSysLocale;
+  getDeviceProductLanguage = deviceSettingsCommands.getDeviceProductLanguage;
+  getDeviceProductCountry = deviceSettingsCommands.getDeviceProductCountry;
+  getDeviceProductLocale = deviceSettingsCommands.getDeviceProductLocale;
+  getModel = deviceSettingsCommands.getModel;
+  getManufacturer = deviceSettingsCommands.getManufacturer;
+  getScreenSize = deviceSettingsCommands.getScreenSize;
+  getScreenDensity = deviceSettingsCommands.getScreenDensity;
+  setHttpProxy = deviceSettingsCommands.setHttpProxy;
+  deleteHttpProxy = deviceSettingsCommands.deleteHttpProxy;
+  setSetting = deviceSettingsCommands.setSetting;
+  getSetting = deviceSettingsCommands.getSetting;
+  getTimeZone = deviceSettingsCommands.getTimeZone;
+  getPlatformVersion = deviceSettingsCommands.getPlatformVersion;
+  getLocationProviders = deviceSettingsCommands.getLocationProviders;
+  toggleGPSLocationProvider = deviceSettingsCommands.toggleGPSLocationProvider;
+  setHiddenApiPolicy = deviceSettingsCommands.setHiddenApiPolicy;
+  setDefaultHiddenApiPolicy = deviceSettingsCommands.setDefaultHiddenApiPolicy;
+  getDeviceLanguage = deviceSettingsCommands.getDeviceLanguage;
+  getDeviceCountry = deviceSettingsCommands.getDeviceCountry;
+  getDeviceLocale = deviceSettingsCommands.getDeviceLocale;
+  ensureCurrentLocale = deviceSettingsCommands.ensureCurrentLocale;
+  setWifiState = deviceSettingsCommands.setWifiState;
+  setDataState = deviceSettingsCommands.setDataState;
+  getDeviceIdleWhitelist = deviceSettingsCommands.getDeviceIdleWhitelist;
+  addToDeviceIdleWhitelist = deviceSettingsCommands.addToDeviceIdleWhitelist;
+  isAirplaneModeOn = deviceSettingsCommands.isAirplaneModeOn;
+  setAirplaneMode = deviceSettingsCommands.setAirplaneMode;
+  setBluetoothOn = deviceSettingsCommands.setBluetoothOn;
+  setNfcOn = deviceSettingsCommands.setNfcOn;
+  broadcastAirplaneMode = deviceSettingsCommands.broadcastAirplaneMode;
+  isWifiOn = deviceSettingsCommands.isWifiOn;
+  isDataOn = deviceSettingsCommands.isDataOn;
+  isAnimationOn = deviceSettingsCommands.isAnimationOn;
+  setAnimationScale = deviceSettingsCommands.setAnimationScale;
+  getScreenOrientation = deviceSettingsCommands.getScreenOrientation;
 }
