+++ lib/tools/adb-commands.js
@@ -1,7 +1,6 @@
 import log from '../logger.js';
 import {
-  getIMEListFromOutput, getSurfaceOrientation, extractMatchingPermissions,
-  parseLaunchableActivityNames, matchComponentName,
+  extractMatchingPermissions, parseLaunchableActivityNames, matchComponentName,
 } from '../helpers.js';
 import path from 'path';
 import _ from 'lodash';
@@ -19,16 +18,6 @@ const IGNORED_PERM_ERRORS = [
   /Unknown permission/i,
 ];
 const MAX_PGREP_PATTERN_LEN = 15;
-const HIDDEN_API_POLICY_KEYS = [
-  'hidden_api_policy_pre_p_apps',
-  'hidden_api_policy_p_apps',
-  'hidden_api_policy'
-];
-const ANIMATION_SCALE_KEYS = [
-  'animator_duration_scale',
-  'transition_animation_scale',
-  'window_animation_scale'
-];
 const PID_COLUMN_TITLE = 'PID';
 const PROCESS_NAME_COLUMN_TITLE = 'NAME';
 const PS_TITLE_PATTERN = new RegExp(`^(.*\\b${PID_COLUMN_TITLE}\\b.*\\b${PROCESS_NAME_COLUMN_TITLE}\\b.*)$`, 'm');
@@ -174,25 +163,6 @@ export async function getApiLevel () {
     }
   }
   return /** @type {number} */(this._apiLevel);
-}
-
-/**
- * Retrieve the platform version of the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} The platform version as a string, for example '5.0' for
- * Android Lollipop.
- */
-export async function getPlatformVersion () {
-  log.info('Getting device platform version');
-  try {
-    return await this.getDeviceProperty('ro.build.version.release');
-  } catch (e) {
-    throw new Error(
-      `Error getting device platform version. ` +
-      `Original error: ${(/**@type {Error} */ (e)).message}`
-    );
-  }
 }
 
 /**
@@ -483,118 +453,6 @@ export async function getReqPermissions (pkg, cmdOutput = null) {
 }
 
 /**
- * Retrieve the list of location providers for the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<string[]>} The list of available location providers or an empty list.
- */
-export async function getLocationProviders () {
-  if (await this.getApiLevel() < 31) {
-    // https://stackoverflow.com/questions/70939503/settings-secure-location-providers-allowed-returns-null-in-android-12
-    const stdout = await this.getSetting('secure', 'location_providers_allowed');
-    return stdout.trim().split(',')
-      .map((p) => p.trim())
-      .filter(Boolean);
-  }
-
-  // To emulate the legacy behavior
-  return _.includes(await this.shell(['cmd', 'location', 'is-location-enabled']), 'true')
-    ? ['gps']
-    : [];
-}
-
-/**
- * Toggle the state of GPS location provider.
- *
- * @this {import('../adb.js').ADB}
- * @param {boolean} enabled - Whether to enable (true) or disable (false) the GPS provider.
- */
-export async function toggleGPSLocationProvider (enabled) {
-  if (await this.getApiLevel() < 31) {
-    // https://stackoverflow.com/questions/70939503/settings-secure-location-providers-allowed-returns-null-in-android-12
-    await this.setSetting('secure', 'location_providers_allowed', `${enabled ? '+' : '-'}gps`);
-    return;
-  }
-  await this.shell(['cmd', 'location', 'set-location-enabled', enabled ? 'true' : 'false']);
-}
-
-/**
- * Decorates an exception message with a solution link
- *
- * @param {Error} e The error object to be decorated
- * @returns {Error} Either the same error or the decorated one
- */
-function decorateWriteSecureSettingsException (e) {
-  if (_.includes(e.message, 'requires:android.permission.WRITE_SECURE_SETTINGS')) {
-    e.message = `Check https://github.com/appium/appium/issues/13802 for throubleshooting. ${e.message}`;
-  }
-  return e;
-}
-
-/**
- * Set hidden api policy to manage access to non-SDK APIs.
- * https://developer.android.com/preview/restrictions-non-sdk-interfaces
- *
- * @this {import('../adb.js').ADB}
- * @param {number|string} value - The API enforcement policy.
- *     For Android P
- *     0: Disable non-SDK API usage detection. This will also disable logging, and also break the strict mode API,
- *        detectNonSdkApiUsage(). Not recommended.
- *     1: "Just warn" - permit access to all non-SDK APIs, but keep warnings in the log.
- *        The strict mode API will keep working.
- *     2: Disallow usage of dark grey and black listed APIs.
- *     3: Disallow usage of blacklisted APIs, but allow usage of dark grey listed APIs.
- *
- *     For Android Q
- *     https://developer.android.com/preview/non-sdk-q#enable-non-sdk-access
- *     0: Disable all detection of non-SDK interfaces. Using this setting disables all log messages for non-SDK interface usage
- *        and prevents you from testing your app using the StrictMode API. This setting is not recommended.
- *     1: Enable access to all non-SDK interfaces, but print log messages with warnings for any non-SDK interface usage.
- *        Using this setting also allows you to test your app using the StrictMode API.
- *     2: Disallow usage of non-SDK interfaces that belong to either the black list
- *        or to a restricted greylist for your target API level.
- *
- * @param {boolean} [ignoreError=false] Whether to ignore an exception in 'adb shell settings put global' command
- * @throws {error} If there was an error and ignoreError was true while executing 'adb shell settings put global'
- *                 command on the device under test.
- */
-export async function setHiddenApiPolicy (value, ignoreError = false) {
-  try {
-    await this.shell(HIDDEN_API_POLICY_KEYS.map((k) => `settings put global ${k} ${value}`).join(';'));
-  } catch (e) {
-    const err = /** @type {Error} */ (e);
-    if (!ignoreError) {
-      throw decorateWriteSecureSettingsException(err);
-    }
-    log.info(
-      `Failed to set setting keys '${HIDDEN_API_POLICY_KEYS}' to '${value}'. ` +
-      `Original error: ${err.message}`
-    );
-  }
-}
-
-/**
- * Reset access to non-SDK APIs to its default setting.
- * https://developer.android.com/preview/restrictions-non-sdk-interfaces
- *
- * @this {import('../adb.js').ADB}
- * @param {boolean} [ignoreError=false] Whether to ignore an exception in 'adb shell settings delete global' command
- * @throws {error} If there was an error and ignoreError was true while executing 'adb shell settings delete global'
- *                 command on the device under test.
- */
-export async function setDefaultHiddenApiPolicy (ignoreError = false) {
-  try {
-    await this.shell(HIDDEN_API_POLICY_KEYS.map((k) => `settings delete global ${k}`).join(';'));
-  } catch (e) {
-    const err = /** @type {Error} */ (e);
-    if (!ignoreError) {
-      throw decorateWriteSecureSettingsException(err);
-    }
-    log.info(`Failed to delete keys '${HIDDEN_API_POLICY_KEYS}'. Original error: ${err.message}`);
-  }
-}
-
-/**
  * Stop the particular package if it is running and clears its application data.
  *
  * @this {import('../adb.js').ADB}
@@ -608,128 +466,6 @@ export async function stopAndClear (pkg) {
     const err = /** @type {Error} */ (e);
     throw new Error(`Cannot stop and clear ${pkg}. Original error: ${err.message}`);
   }
-}
-
-/**
- * Retrieve the list of available input methods (IMEs) for the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<string[]>} The list of IME names or an empty list.
- */
-export async function availableIMEs () {
-  try {
-    return getIMEListFromOutput(await this.shell(['ime', 'list', '-a']));
-  } catch (e) {
-    const err = /** @type {Error} */ (e);
-    throw new Error(`Error getting available IME's. Original error: ${err.message}`);
-  }
-}
-
-/**
- * Retrieve the list of enabled input methods (IMEs) for the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<string[]>} The list of enabled IME names or an empty list.
- */
-export async function enabledIMEs () {
-  try {
-    return getIMEListFromOutput(await this.shell(['ime', 'list']));
-  } catch (e) {
-    const err = /** @type {Error} */ (e);
-    throw new Error(`Error getting enabled IME's. Original error: ${err.message}`);
-  }
-}
-
-/**
- * Enable the particular input method on the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @param {string} imeId - One of existing IME ids.
- */
-export async function enableIME (imeId) {
-  await this.shell(['ime', 'enable', imeId]);
-}
-
-/**
- * Disable the particular input method on the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @param {string} imeId - One of existing IME ids.
- */
-export async function disableIME (imeId) {
-  await this.shell(['ime', 'disable', imeId]);
-}
-
-/**
- * Set the particular input method on the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @param {string} imeId - One of existing IME ids.
- */
-export async function setIME (imeId) {
-  await this.shell(['ime', 'set', imeId]);
-}
-
-/**
- * Get the default input method on the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<string|null>} The name of the default input method
- */
-export async function defaultIME () {
-  try {
-    let engine = await this.getSetting('secure', 'default_input_method');
-    if (engine === 'null') {
-      return null;
-    }
-    return engine.trim();
-  } catch (e) {
-    const err = /** @type {Error} */ (e);
-    throw new Error(`Error getting default IME. Original error: ${err.message}`);
-  }
-}
-
-/**
- * Send the particular keycode to the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @param {string|number} keycode - The actual key code to be sent.
- */
-export async function keyevent (keycode) {
-  // keycode must be an int.
-  const code = parseInt(`${keycode}`, 10);
-  await this.shell(['input', 'keyevent', `${code}`]);
-}
-
-/**
- * Send the particular text or a number to the device under test.
- * The text gets properly escaped before being passed to ADB.
- * Noop if the text is empty.
- *
- * @this {import('../adb.js').ADB}
- * @param {string|number} text - The actual text to be sent.
- * @throws {Error} If it is impossible to escape the given string
- */
-export async function inputText (text) {
-  if (text === '') {
-    return;
-  }
-
-  const originalStr = `${text}`;
-  const escapedText = originalStr.replace(/\$/g, '\\$').replace(/ /g, '%s');
-  let args = ['input', 'text', originalStr];
-  // https://stackoverflow.com/questions/25791423/adb-shell-input-text-does-not-take-ampersand-character/25791498
-  const adbInputEscapePattern = /[()<>|;&*\\~^"']/g;
-  if (escapedText !== originalStr || adbInputEscapePattern.test(originalStr)) {
-    if (_.every(['"', `'`], (c) => originalStr.includes(c))) {
-      throw new Error(
-        `Did not know how to escape a string that contains both types of quotes (" and ')`
-      );
-    }
-    const q = originalStr.includes('"') ? `'` : '"';
-    args = [`input text ${q}${escapedText}${q}`];
-  }
-  await this.shell(args);
 }
 
 /**
@@ -808,17 +544,6 @@ export function getAdbPath () {
 }
 
 /**
- * Retrieve current screen orientation of the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<number?>} The current orientation encoded as an integer number.
- */
-export async function getScreenOrientation () {
-  let stdout = await this.shell(['dumpsys', 'input']);
-  return getSurfaceOrientation(stdout);
-}
-
-/**
  * Send an arbitrary Telnet command to the device under test.
  *
  * @this {import('../adb.js').ADB}
@@ -827,162 +552,6 @@ export async function getScreenOrientation () {
  */
 export async function sendTelnetCommand (command) {
   return await this.execEmuConsoleCommand(command, {port: await this.getEmulatorPort()});
-}
-
-/**
- * Check the state of Airplane mode on the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<boolean>} True if Airplane mode is enabled.
- */
-export async function isAirplaneModeOn () {
-  const stdout = await this.getSetting('global', 'airplane_mode_on');
-  return parseInt(stdout, 10) !== 0;
-  // Alternatively for Android 11+:
-  // return (await this.shell(['cmd', 'connectivity', 'airplane-mode'])).stdout.trim() === 'enabled';
-}
-
-/**
- * Change the state of Airplane mode in Settings on the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @param {boolean} on - True to enable the Airplane mode in Settings and false to disable it.
- */
-export async function setAirplaneMode (on) {
-  if (await this.getApiLevel() < 30) {
-    // This requires to call broadcastAirplaneMode afterwards to apply
-    await this.setSetting('global', 'airplane_mode_on', on ? 1 : 0);
-    return;
-  }
-
-  await this.shell(['cmd', 'connectivity', 'airplane-mode', on ? 'enable' : 'disable']);
-}
-
-/**
- * Change the state of the bluetooth service on the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @param {boolean} on - True to enable bluetooth service and false to disable it.
- */
-export async function setBluetoothOn (on) {
-  if (await this.getApiLevel() < 30) {
-    throw new Error('Changing of the bluetooth state is not supported on your device');
-  }
-
-  await this.shell(['cmd', 'bluetooth_manager', on ? 'enable' : 'disable']);
-}
-
-/**
- * Change the state of the NFC service on the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @param {boolean} on - True to enable NFC service and false to disable it.
- * @throws {Error} If there was an error while changing the service state
- */
-export async function setNfcOn (on) {
-  const {stdout, stderr} = await this.shell(['svc', 'nfc', on ? 'enable' : 'disable'], {
-    outputFormat: 'full'
-  });
-  const output = stderr || stdout;
-  log.debug(output);
-  if (output.includes('null NfcAdapter')) {
-    throw new Error(
-      `Cannot turn ${on ? 'on' : 'off'} the NFC adapter. Does the device under test have it?`
-    );
-  }
-}
-
-/**
- * Broadcast the state of Airplane mode on the device under test.
- * This method should be called after {@link #setAirplaneMode}, otherwise
- * the mode change is not going to be applied for the device.
- * ! This API requires root since Android API 24. Since API 30
- * there is a dedicated adb command to change airplane mode state, which
- * does not require to call this one afterwards.
- *
- * @this {import('../adb.js').ADB}
- * @param {boolean} on - True to broadcast enable and false to broadcast disable.
- */
-export async function broadcastAirplaneMode (on) {
-  const args = [
-    'am', 'broadcast',
-    '-a', 'android.intent.action.AIRPLANE_MODE',
-    '--ez', 'state', on ? 'true' : 'false',
-  ];
-  try {
-    await this.shell(args);
-  } catch (e) {
-    const err = /** @type {import('teen_process').ExecError} */(e);
-    // https://github.com/appium/appium/issues/17422
-    if (_.includes(err.stderr, 'SecurityException')) {
-      try {
-        await this.shell(args, {privileged: true});
-        return;
-      } catch {}
-    }
-    throw err;
-  }
-}
-
-/**
- * Check the state of WiFi on the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<boolean>} True if WiFi is enabled.
- */
-export async function isWifiOn () {
-  const stdout = await this.getSetting('global', 'wifi_on');
-  return (parseInt(stdout, 10) !== 0);
-  // Alternative for Android 11+:
-  // return (await this.shell(['cmd', 'wifi', 'status']).stdout.includes('Wifi is enabled'));
-}
-
-/**
- * Check the state of Data transfer on the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<boolean>} True if Data transfer is enabled.
- */
-export async function isDataOn () {
-  const stdout = await this.getSetting('global', 'mobile_data');
-  return (parseInt(stdout, 10) !== 0);
-}
-
-/**
- * Check the state of animation on the device under test below:
- *   - animator_duration_scale
- *   - transition_animation_scale
- *   - window_animation_scale
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<boolean>} True if at least one of animation scale settings
- *                   is not equal to '0.0'.
- */
-export async function isAnimationOn () {
-  return (await B.all(ANIMATION_SCALE_KEYS.map(
-    async (k) => (await this.getSetting('global', k)) !== '0.0'))
-  ).includes(true);
-}
-
-/**
- * Set animation scale with the given value via adb shell settings command.
- *   - animator_duration_scale
- *   - transition_animation_scale
- *   - window_animation_scale
- * API level 24 and newer OS versions may change the animation, at least emulators are so.
- * API level 28+ real devices checked this worked, but we haven't checked older ones
- * with real devices.
- *
- * @this {import('../adb.js').ADB}
- * @param {number} value Animation scale value (int or float) to set.
- *                       The minimum value of zero disables animations.
- *                       By increasing the value, animations become slower.
- *                       '1' is the system default animation scale.
- * @return {Promise<void>}
- * @throws {Error} If the adb setting command raises an exception.
- */
-export async function setAnimationScale (value) {
-  await B.all(ANIMATION_SCALE_KEYS.map((k) => this.setSetting('global', k, value)));
 }
 
 /**
@@ -1494,210 +1063,6 @@ export async function broadcast (intent) {
 }
 
 /**
- * Get the particular property of the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @param {string} property - The name of the property. This name should
- *                            be known to _adb shell getprop_ tool.
- *
- * @return {Promise<string>} The value of the given property.
- */
-export async function getDeviceProperty (property) {
-  let stdout = await this.shell(['getprop', property]);
-  let val = stdout.trim();
-  log.debug(`Current device property '${property}': ${val}`);
-  return val;
-}
-
-/**
- * Set the particular property of the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @param {string} prop - The name of the property. This name should
- *                            be known to _adb shell setprop_ tool.
- * @param {string} val - The new property value.
- * @param {import('./types').SetPropOpts} [opts={}]
- *
- * @throws {error} If _setprop_ utility fails to change property value.
- */
-export async function setDeviceProperty (prop, val, opts = {}) {
-  const {privileged = true} = opts;
-  log.debug(`Setting device property '${prop}' to '${val}'`);
-  await this.shell(['setprop', prop, val], {
-    privileged,
-  });
-}
-
-/**
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} Current system language on the device under test.
- */
-export async function getDeviceSysLanguage () {
-  return await this.getDeviceProperty('persist.sys.language');
-}
-
-/**
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} Current country name on the device under test.
- */
-export async function getDeviceSysCountry () {
-  return await this.getDeviceProperty('persist.sys.country');
-}
-
-/**
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} Current system locale name on the device under test.
- */
-export async function getDeviceSysLocale () {
-  return await this.getDeviceProperty('persist.sys.locale');
-}
-
-/**
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} Current product language name on the device under test.
- */
-export async function getDeviceProductLanguage () {
-  return await this.getDeviceProperty('ro.product.locale.language');
-}
-
-/**
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} Current product country name on the device under test.
- */
-export async function getDeviceProductCountry () {
-  return await this.getDeviceProperty('ro.product.locale.region');
-}
-
-/**
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} Current product locale name on the device under test.
- */
-export async function getDeviceProductLocale () {
-  return await this.getDeviceProperty('ro.product.locale');
-}
-
-/**
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} The model name of the device under test.
- */
-export async function getModel () {
-  return await this.getDeviceProperty('ro.product.model');
-}
-
-/**
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} The manufacturer name of the device under test.
- */
-export async function getManufacturer () {
-  return await this.getDeviceProperty('ro.product.manufacturer');
-}
-
-/**
- * Get the current screen size.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<string?>} Device screen size as string in format 'WxH' or
- * _null_ if it cannot be determined.
- */
-export async function getScreenSize () {
-  let stdout = await this.shell(['wm', 'size']);
-  let size = new RegExp(/Physical size: ([^\r?\n]+)*/g).exec(stdout);
-  if (size && size.length >= 2) {
-    return size[1].trim();
-  }
-  return null;
-}
-
-/**
- * Get the current screen density in dpi
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<number?>} Device screen density as a number or _null_ if it
- * cannot be determined
- */
-export async function getScreenDensity () {
-  let stdout = await this.shell(['wm', 'density']);
-  let density = new RegExp(/Physical density: ([^\r?\n]+)*/g).exec(stdout);
-  if (density && density.length >= 2) {
-    let densityNumber = parseInt(density[1].trim(), 10);
-    return isNaN(densityNumber) ? null : densityNumber;
-  }
-  return null;
-}
-
-/**
- * Setup HTTP proxy in device global settings.
- * Read https://android.googlesource.com/platform/frameworks/base/+/android-9.0.0_r21/core/java/android/provider/Settings.java for each property
- *
- * @this {import('../adb.js').ADB}
- * @param {string} proxyHost - The host name of the proxy.
- * @param {string|number} proxyPort - The port number to be set.
- */
-export async function setHttpProxy (proxyHost, proxyPort) {
-  let proxy = `${proxyHost}:${proxyPort}`;
-  if (_.isUndefined(proxyHost)) {
-    throw new Error(`Call to setHttpProxy method with undefined proxy_host: ${proxy}`);
-  }
-  if (_.isUndefined(proxyPort)) {
-    throw new Error(`Call to setHttpProxy method with undefined proxy_port ${proxy}`);
-  }
-
-  /** @type {[string, string][]} */
-  const httpProxySettins = [
-    ['http_proxy', proxy],
-    ['global_http_proxy_host', proxyHost],
-    ['global_http_proxy_port', `${proxyPort}`]
-  ];
-  for (const [settingKey, settingValue] of httpProxySettins) {
-    await this.setSetting('global', settingKey, settingValue);
-  }
-}
-
-/**
- * Delete HTTP proxy in device global settings.
- * Rebooting the test device is necessary to apply the change.
- * @this {import('../adb.js').ADB}
- */
-export async function deleteHttpProxy () {
-  const httpProxySettins = [
-    'http_proxy',
-    'global_http_proxy_host',
-    'global_http_proxy_port',
-    'global_http_proxy_exclusion_list' // `global_http_proxy_exclusion_list=` was generated by `settings global htto_proxy xxxx`
-  ];
-  for (const setting of httpProxySettins) {
-    await this.shell(['settings', 'delete', 'global', setting]);
-  }
-}
-
-/**
- * Set device property.
- * [android.provider.Settings]{@link https://developer.android.com/reference/android/provider/Settings.html}
- *
- * @this {import('../adb.js').ADB}
- * @param {string} namespace - one of {system, secure, global}, case-insensitive.
- * @param {string} setting - property name.
- * @param {string|number} value - property value.
- * @return {Promise<string>} command output.
- */
-export async function setSetting (namespace, setting, value) {
-  return await this.shell(['settings', 'put', namespace, setting, `${value}`]);
-}
-
-/**
- * Get device property.
- * [android.provider.Settings]{@link https://developer.android.com/reference/android/provider/Settings.html}
- *
- * @this {import('../adb.js').ADB}
- * @param {string} namespace - one of {system, secure, global}, case-insensitive.
- * @param {string} setting - property name.
- * @return {Promise<string>} property value.
- */
-export async function getSetting (namespace, setting) {
-  return await this.shell(['settings', 'get', namespace, setting]);
-}
-
-/**
  * Retrieve the `adb bugreport` command output. This
  * operation may take up to several minutes.
  *
@@ -1747,50 +1112,6 @@ export function screenrecord (destination, options = {}) {
   ];
   log.debug(`Building screenrecord process with the command line: adb ${util.quote(fullCmd)}`);
   return new SubProcess(this.executable.path, fullCmd);
-}
-
-/**
- * Executes the given function with the given input method context
- * and then restores the IME to the original value
- *
- * @this {import('../adb.js').ADB}
- * @param {string} ime - Valid IME identifier
- * @param {Function} fn - Function to execute
- * @returns {Promise<any>} The result of the given function
- */
-export async function runInImeContext (ime, fn) {
-  const originalIme = await this.defaultIME();
-  if (originalIme === ime) {
-    log.debug(`The original IME is the same as '${ime}'. There is no need to reset it`);
-  } else {
-    await this.enableIME(ime);
-    await this.setIME(ime);
-    // https://github.com/appium/appium/issues/15943
-    await B.delay(500);
-  }
-  try {
-    return await fn();
-  } finally {
-    if (originalIme && originalIme !== ime) {
-      await this.setIME(originalIme);
-    }
-  }
-}
-
-/**
- * Get tz database time zone formatted timezone
- *
- * @this {import('../adb.js').ADB}
- * @returns {Promise<string>} TZ database Time Zones format
- * @throws {Error} If any exception is reported by adb shell.
- */
-export async function getTimeZone () {
-  log.debug('Getting current timezone');
-  try {
-    return await this.getDeviceProperty('persist.sys.timezone');
-  } catch (e) {
-    throw new Error(`Error getting timezone. Original error: ${(/** @type {Error} */ (e)).message}`);
-  }
 }
 
 /**
@@ -1870,49 +1191,6 @@ export async function isIncrementalInstallSupported () {
 }
 
 /**
- * Retrieves the list of packages from Doze whitelist on Android 8+
- *
- * @this {import('../adb.js').ADB}
- * @returns {Promise<string[]>} The list of whitelisted packages. An example output:
- * system,com.android.shell,2000
- * system,com.google.android.cellbroadcastreceiver,10143
- * user,io.appium.settings,10157
- */
-export async function getDeviceIdleWhitelist () {
-  if (await this.getApiLevel() < 23) {
-    // Doze mode has only been added since Android 6
-    return [];
-  }
-
-  log.info('Listing packages in Doze whitelist');
-  const output = await this.shell(['dumpsys', 'deviceidle', 'whitelist']);
-  return _.trim(output).split(/\n/)
-    .map((line) => _.trim(line))
-    .filter(Boolean);
-}
-
-/**
- * Adds an existing package(s) into the Doze whitelist on Android 8+
- *
- * @this {import('../adb.js').ADB}
- * @param  {...string} packages One or more packages to add. If the package
- * already exists in the whitelist then it is only going to be added once.
- * If the package with the given name is not installed/not known then an error
- * will be thrown.
- * @returns {Promise<boolean>} `true` if the command to add package(s) has been executed
- */
-export async function addToDeviceIdleWhitelist (...packages) {
-  if (_.isEmpty(packages) || await this.getApiLevel() < 23) {
-    // Doze mode has only been added since Android 6
-    return false;
-  }
-
-  log.info(`Adding ${util.pluralize('package', packages.length)} ${JSON.stringify(packages)} to Doze whitelist`);
-  await this.shellChunks((pkg) => ['dumpsys', 'deviceidle', 'whitelist', `+${pkg}`], packages);
-  return true;
-}
-
-/**
  * Takes a screenshot of the given display or the default display.
  *
  * @this {import('../adb.js').ADB}
@@ -1946,48 +1224,6 @@ export async function takeScreenshot (displayId) {
     throw new Error(`Screenshot of the ${displayDescr} returned no data`);
   }
   return stdout;
-}
-
-/**
- * Change the state of WiFi on the device under test.
- * Only works for real devices since API 30
- *
- * @this {import('../adb.js').ADB}
- * @param {boolean} on - True to enable and false to disable it.
- * @param {boolean} [isEmulator=false] - Set it to true if the device under test
- *                                       is an emulator rather than a real device.
- */
-export async function setWifiState (on, isEmulator = false) {
-  if (isEmulator) {
-    // The svc command does not require to be root since API 26
-    await this.shell(['svc', 'wifi', on ? 'enable' : 'disable'], {
-      privileged: await this.getApiLevel() < 26,
-    });
-    return;
-  }
-
-  await this.shell(['cmd', '-w', 'wifi', 'set-wifi-enabled', on ? 'enabled' : 'disabled']);
-}
-
-/**
- * Change the state of Data transfer on the device under test.
- * Only works for real devices since API 30
- *
- * @this {import('../adb.js').ADB}
- * @param {boolean} on - True to enable and false to disable it.
- * @param {boolean} [isEmulator=false] - Set it to true if the device under test
- *                                       is an emulator rather than a real device.
- */
-export async function setDataState (on, isEmulator = false) {
-  if (isEmulator) {
-    // The svc command does not require to be root since API 26
-    await this.shell(['svc', 'data', on ? 'enable' : 'disable'], {
-      privileged: await this.getApiLevel() < 26,
-    });
-    return;
-  }
-
-  await this.shell(['cmd', 'phone', 'data', on ? 'enable' : 'disable']);
 }
 
 /**
