+++ lib/tools/apk-utils.js
@@ -8,7 +8,7 @@ import { exec } from 'teen_process';
 import log from '../logger.js';
 import path from 'path';
 import _ from 'lodash';
-import { retryInterval, waitForCondition } from 'asyncbox';
+import { waitForCondition } from 'asyncbox';
 import { fs, util, mkdirp, timing } from '@appium/support';
 import * as semver from 'semver';
 import os from 'os';
@@ -854,102 +854,6 @@ export async function extractStringsFromApk (
   await mkdirp(outRoot);
   await fs.writeFile(localPath, JSON.stringify(apkStrings, null, 2), 'utf-8');
   return {apkStrings, localPath};
-}
-
-/**
- * Get the language name of the device under test.
- *
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} The name of device language.
- */
-export async function getDeviceLanguage () {
-  return await this.getApiLevel() < 23
-    ? (await this.getDeviceSysLanguage() || await this.getDeviceProductLanguage())
-    : (await this.getDeviceLocale()).split('-')[0];
-}
-
-/**
- * Get the country name of the device under test.
- *
- * @summary Could only be used for Android API < 23
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} The name of device country.
- */
-export async function getDeviceCountry () {
-  return await this.getDeviceSysCountry() || await this.getDeviceProductCountry();
-}
-
-/**
- * Get the locale name of the device under test.
- *
- * @summary Could only be used for Android API >= 23
- * @this {import('../adb.js').ADB}
- * @return {Promise<string>} The name of device locale.
- */
-export async function getDeviceLocale () {
-  return await this.getDeviceSysLocale() || await this.getDeviceProductLocale();
-}
-
-/**
- * Make sure current device locale is expected or not.
- *
- * @this {import('../adb.js').ADB}
- * @privateRemarks FIXME: language or country is required
- * @param {string} [language] - Language. The language field is case insensitive, but Locale always canonicalizes to lower case.
- * @param {string} [country] - Country. The language field is case insensitive, but Locale always canonicalizes to lower case.
- * @param {string} [script] - Script. The script field is case insensitive but Locale always canonicalizes to title case.
- *
- * @return {Promise<boolean>} If current locale is language and country as arguments, return true.
- */
-export async function ensureCurrentLocale (language, country, script) {
-  const hasLanguage = _.isString(language);
-  const hasCountry = _.isString(country);
-  if (!hasLanguage && !hasCountry) {
-    log.warn('ensureCurrentLocale requires language or country');
-    return false;
-  }
-
-  const lcLanguage = (language || '').toLowerCase();
-  const lcCountry = (country || '').toLowerCase();
-  const apiLevel = await this.getApiLevel();
-  return /** @type {boolean} */ (await retryInterval(5, 1000, async () => {
-    if (apiLevel < 23) {
-      log.debug(`Requested locale: ${lcLanguage}-${lcCountry}`);
-      let actualLanguage;
-      if (hasLanguage) {
-        actualLanguage = (await this.getDeviceLanguage()).toLowerCase();
-        log.debug(`Actual language: ${actualLanguage}`);
-        if (!hasCountry && lcLanguage === actualLanguage) {
-          return true;
-        }
-      }
-      let actualCountry;
-      if (hasCountry) {
-        actualCountry = (await this.getDeviceCountry()).toLowerCase();
-        log.debug(`Actual country: ${actualCountry}`);
-        if (!hasLanguage && lcCountry === actualCountry) {
-          return true;
-        }
-      }
-      return lcLanguage === actualLanguage && lcCountry === actualCountry;
-    }
-    const actualLocale = (await this.getDeviceLocale()).toLowerCase();
-    // zh-hans-cn : zh-cn
-    const expectedLocale = script
-      ? `${lcLanguage}-${script.toLowerCase()}-${lcCountry}`
-      : `${lcLanguage}-${lcCountry}`;
-    log.debug(`Requested locale: ${expectedLocale}. Actual locale: '${actualLocale}'`);
-    const languagePattern = `^${_.escapeRegExp(lcLanguage)}-${script ? (_.escapeRegExp(script) + '-') : ''}`;
-    const checkLocalePattern = (/** @type {string} */ p) => new RegExp(p, 'i').test(actualLocale);
-    if (hasLanguage && !hasCountry) {
-      return checkLocalePattern(languagePattern);
-    }
-    const countryPattern = `${script ? ('-' + _.escapeRegExp(script)) : ''}-${_.escapeRegExp(lcCountry)}$`;
-    if (!hasLanguage && hasCountry) {
-      return checkLocalePattern(countryPattern);
-    }
-    return [languagePattern, countryPattern].every(checkLocalePattern);
-  }));
 }
 
 /**
